Implemented one viable solution for the validator api.
While looking at the code I found some inconsistencies.
You where passing the validators to the mapper. There was no real solution for using them
as the interface statet a boolean but you were clearly expecting an exception.
I changed the Validator interface to void and let it throw an exception on error, like all
other classes you've build. This way you can kill the TypeMap.validate() method as it is of no use
anymore. The way it is implemented now the typemap does not need to know anything about validators.
It just exposes two different sets of unmapped properties that can be checked against.

This solution is not perfect and can be discussed but I think its the fastest and "cleanest" way to
work a little against the hard coupling in all the classes.

Suggestions welcome :)