Add support for mapping to SortedSets destinations.
Previously the CollectionConverter was only discriminating between Lists and
Sets, instanciating an ArrayList if the destination type was a List and a
HashSet if the destination type was a Set. This failed when the destination
type was a SortedSet (HashSet is not a SortedSet).

Rather than change the HashSet to a TereSet for all Sets, a separate check has
been added for SortedSets so that the more performant HashMap is still used
generally and a TreeSet is used only when we know we need a SortedSet.

Two test cases where added. One is to test that the SortedSet mapping works.
The other is to test that HashSets are still used when a SortedSet is not
requested. This second test may not be necessary if we don't care that much
about the implementation backing the mapped property.