Implement TypeMap inheritance
Currently, when ModelMapper decide which TypeMap to use,
it was strongly depends on the actually type of source and
destination.
But sometimes we want to define TypeMap with interface
or parent class to map there model, or configure a TypeMap to inherit
a base TypeMap.
So we introduce TypeMap.include and TypeMap.includeBase so that we can
have TypeMap inheritance.

Resolved: #13 #130 #132 #170 #202