Improved building TypeMap with ValueReader
Currently, TypeMap will always be cached into TypeMapStore,
even it's had some properties still not be mapped properly.
And it would cause problem when handing source type that is
dynamic type and need a ValueReader to read its value, like
Json, Map, DB ResultSet or Record.

So I had some improvement for these not handling well stuff.
1. ValueReader with new method getAccessor
So when encounter flat dynamic type, like jooq Record, the
accessors get from ValueReader will not depends on the value,
but depends on the metadata.

2. Only cache when TypeMap is completely mapping
When handling some hierarchical dynamic type, we cannot prevent
that the value to not be null. We must not cache for that
situation, so that we can create a new TypeMap next time and next
time, until we got a completely mapped TypeMap and we will cache it.

3. Re-implement how TypeMap determine un-mapped properties
Current implementation of unmapped properties is only check the
top-layer fields. But it's not true about a completely mapped
TypeMap.

Closes: #168