Returns null instead of throw error for ProxyClass
We try to proxy all properties when user build an explicit map,
but we cannot proxy some class with cglib like EnumSet. But for
some case, we don't need to proxy all properties, so we try to
return null instead of throw error instantly. And we throw errors
when we got NullPointerException.

Resolved: #260